// src/__generated__/types/outlay-row-request.ts
export type OutlayRowRequest = {
    /**
     * @type integer | undefined, int64
    */
    equipmentCosts?: number;
    /**
     * @type integer | undefined, int64
    */
    estimatedProfit?: number;
    /**
     * @type integer | undefined, int64
    */
    machineOperatorSalary?: number;
    /**
     * @type integer | undefined, int64
    */
    mainCosts?: number;
    /**
     * @type integer | undefined, int64
    */
    materials?: number;
    /**
     * @type integer | undefined, int64
    */
    mimExploitation?: number;
    /**
     * @type integer | undefined, int64
    */
    overheads?: number;
    /**
     * @type integer | undefined, int64
    */
    parentId?: number;
    /**
     * @type string | undefined
    */
    rowName?: string;
    /**
     * @type integer | undefined, int64
    */
    salary?: number;
    /**
     * @type integer | undefined, int64
    */
    supportCosts?: number;
};